move cell editors under \editors
